Move TaskGroups out to separate file in src/Query/
