+++ src/Query/TaskGroups.ts
@@ -1,0 +1,76 @@
+import { GroupHeadings, IntermediateTaskGroups } from '../GroupDetails';
+import type { Grouping } from '../Query';
+import type { Task } from '../Task';
+import { TaskGroup } from './TaskGroup';
+
+/**
+ * TaskGroup stores all the groups of tasks generated by any 'group by'
+ * instructions in the task block.
+ */
+export class TaskGroups {
+    private _groups: TaskGroup[] = new Array<TaskGroup>();
+
+    /**
+     * Constructor for TaskGroups
+     * @param {Grouping[]} groups - 0 or more Grouping values,
+     *                              1 per 'group by' line in the task query block
+     * @param {Task[]} tasks - 0 more more Task objects, with all the tasks
+     *                         matching the query, already in sort order
+     */
+    constructor(groups: Grouping[], tasks: Task[]) {
+        const initialGroups = new IntermediateTaskGroups(groups, tasks);
+        this.addTasks(initialGroups);
+    }
+
+    /**
+     * All the tasks matching the query, grouped together, and in the order
+     * that they should be displayed.
+     */
+    public get groups(): TaskGroup[] {
+        return this._groups;
+    }
+
+    /**
+     * The total number of tasks matching the query.
+     */
+    public totalTasksCount() {
+        let totalTasksCount = 0;
+        for (const group of this.groups) {
+            totalTasksCount += group.tasks.length;
+        }
+        return totalTasksCount;
+    }
+
+    /**
+     * A human-readable representation of all the task groups.
+     *
+     * Note that this is used in snapshot testing, so if the format is
+     * changed, the snapshots will need to be updated.
+     */
+    public toString(): string {
+        let output = '';
+        for (const taskGroup of this.groups) {
+            output += taskGroup.toString();
+            output += '\n---\n';
+        }
+        const totalTasksCount = this.totalTasksCount();
+        output += `\n${totalTasksCount} tasks\n`;
+        return output;
+    }
+
+    private addTasks(initialGroups: IntermediateTaskGroups) {
+        // Get the headings
+        const grouper = new GroupHeadings(initialGroups.groups);
+
+        // Build a container of all the groups
+        for (const [groups, tasks] of initialGroups.groups) {
+            const groupHeadings = grouper.getHeadingsForTaskGroup(groups);
+            const taskGroup = new TaskGroup(groups, groupHeadings, tasks);
+            this.add(taskGroup);
+        }
+    }
+
+    private add(taskGroup: TaskGroup) {
+        this._groups.push(taskGroup);
+    }
+}
